+++ docs/guides/changelog.md
@@ -18,6 +18,7 @@ description: |-
 - Add 'advertise_hub_tlocs','all_hubs_are_equal','tloc_list_id', 'ipv4_prefix_list_ids', 'ipv6_prefix_list_ids' and 'preference' to 'sdwan_hub_and_spoke_topology_policy_definition' resource and data source
 - Fix issue with deleted profile parcel resource, [link](https://github.com/CiscoDevNet/terraform-provider-sdwan/issues/204)
 - Add `sdwan_transport_management_vpn_profile_parcel` resource and data source
+- Add `sdwan_system_logging_profile_parcel` resource and data source
 
 ## 0.3.7